Restructure topbeat to make functions available as library
Topbeat has lots of functionality which can be reused in other libraries, especially topbeat. To decouple the functionality from topbeat, this PR restructures Topbeat into the beater part and the library part. This makes is possible to reuse and share the functionality between Topbeat and Metricbeat.

Further steps should be taken later to also better split up the testing and introduce ucfg in topbeat.

This refactoring did not change any functionality or output of topbeat.
