+++ topbeat/system/fileysystem_test.go
@@ -1,0 +1,34 @@
+// +build !integration
+
+package system
+
+import (
+	"os"
+	"runtime"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestFileSystemList(t *testing.T) {
+
+	if runtime.GOOS == "darwin" && os.Getenv("TRAVIS") == "true" {
+		t.Skip("FileSystem test fails on Travis/OSX with i/o error")
+	}
+
+	fss, err := GetFileSystemList()
+
+	assert.Nil(t, err)
+	assert.True(t, (len(fss) > 0))
+
+	for _, fs := range fss {
+
+		stat, err := GetFileSystemStat(fs)
+		assert.NoError(t, err)
+
+		assert.True(t, (stat.Total >= 0))
+		assert.True(t, (stat.Free >= 0))
+		assert.True(t, (stat.Avail >= 0))
+		assert.True(t, (stat.Used >= 0))
+	}
+}
