+++ topbeat/system/process_test.go
@@ -1,72 +1,13 @@
 // +build !integration
 
-package beater
+package system
 
 import (
-	"os"
-	"runtime"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func TestGetSystemLoad(t *testing.T) {
-
-	if runtime.GOOS == "windows" {
-		return //no load data on windows
-	}
-
-	load, err := GetSystemLoad()
-
-	assert.NotNil(t, load)
-	assert.Nil(t, err)
-
-	assert.True(t, (load.Load1 > 0))
-	assert.True(t, (load.Load5 > 0))
-	assert.True(t, (load.Load15 > 0))
-}
-
-func TestGetCpuTimes(t *testing.T) {
-
-	cpu_stat, err := GetCpuTimes()
-
-	assert.NotNil(t, cpu_stat)
-	assert.Nil(t, err)
-
-	assert.True(t, (cpu_stat.User > 0))
-	assert.True(t, (cpu_stat.Sys > 0))
-
-}
-
-func TestGetMemory(t *testing.T) {
-	mem, err := GetMemory()
-
-	assert.NotNil(t, mem)
-	assert.Nil(t, err)
-
-	assert.True(t, (mem.Total > 0))
-	assert.True(t, (mem.Used > 0))
-	assert.True(t, (mem.Free >= 0))
-	assert.True(t, (mem.ActualFree >= 0))
-	assert.True(t, (mem.ActualUsed > 0))
-}
-
-func TestGetSwap(t *testing.T) {
-
-	if runtime.GOOS == "windows" {
-		return //no load data on windows
-	}
-
-	swap, err := GetSwap()
-
-	assert.NotNil(t, swap)
-	assert.Nil(t, err)
-
-	assert.True(t, (swap.Total >= 0))
-	assert.True(t, (swap.Used >= 0))
-	assert.True(t, (swap.Free >= 0))
-}
 
 func TestPids(t *testing.T) {
 	pids, err := Pids()
@@ -109,7 +50,7 @@ func TestGetProcess(t *testing.T) {
 		assert.True(t, (process.Cpu.User >= 0))
 		assert.True(t, (process.Cpu.Sys >= 0))
 
-		assert.True(t, (process.ctime.Unix() <= time.Now().Unix()))
+		assert.True(t, (process.Ctime.Unix() <= time.Now().Unix()))
 
 		// it's enough to get valid data for a single process
 		break
@@ -124,29 +65,6 @@ func TestProcState(t *testing.T) {
 	assert.Equal(t, getProcState('D'), "idle")
 	assert.Equal(t, getProcState('T'), "stopped")
 	assert.Equal(t, getProcState('Z'), "zombie")
-}
-
-func TestFileSystemList(t *testing.T) {
-
-	if runtime.GOOS == "darwin" && os.Getenv("TRAVIS") == "true" {
-		t.Skip("FileSystem test fails on Travis/OSX with i/o error")
-	}
-
-	fss, err := GetFileSystemList()
-
-	assert.Nil(t, err)
-	assert.True(t, (len(fss) > 0))
-
-	for _, fs := range fss {
-
-		stat, err := GetFileSystemStat(fs)
-		assert.NoError(t, err)
-
-		assert.True(t, (stat.Total >= 0))
-		assert.True(t, (stat.Free >= 0))
-		assert.True(t, (stat.Avail >= 0))
-		assert.True(t, (stat.Used >= 0))
-	}
 }
 
 // BenchmarkGetProcess runs a benchmark of the GetProcess method with caching
